# filters: use faster regexp package