--- a/go.mod
+++ b/go.mod
@@ -288,6 +288,7 @@ require (
 	github.com/cloudflare/cloudflare-go v0.27.0
 	github.com/gofrs/flock v0.7.1 // indirect
 	github.com/gogo/status v1.1.0
+	github.com/grafana/regexp v0.0.0-20220202152701-6a046c4caf32
 	github.com/oklog/ulid v1.3.1
 )