… go.sum
… vendor/github.com/grafana/regexp/.gitignore
… vendor/github.com/grafana/regexp/LICENSE
… vendor/github.com/grafana/regexp/README.md
… vendor/github.com/grafana/regexp/backtrack.go
… vendor/github.com/grafana/regexp/exec.go
… vendor/github.com/grafana/regexp/onepass.go
… vendor/github.com/grafana/regexp/regexp.go
… vendor/github.com/grafana/regexp/syntax/compile.go
… vendor/github.com/grafana/regexp/syntax/doc.go
… vendor/github.com/grafana/regexp/syntax/make_perl_groups.pl
… vendor/github.com/grafana/regexp/syntax/op_string.go
… vendor/github.com/grafana/regexp/syntax/parse.go
… vendor/github.com/grafana/regexp/syntax/perl_groups.go
… vendor/github.com/grafana/regexp/syntax/prog.go
… vendor/github.com/grafana/regexp/syntax/regexp.go
… vendor/github.com/grafana/regexp/syntax/simplify.go
… vendor/modules.txt