@bboreham 2022-02-04
> does it work the same for capture and such

Yes, the same regexp engine is used for `Match`, `Find` and `Replace`, so they should see some benefit.
I guess we should have a benchmark for those.